refactor: organize auth components
