+++ frontend/src/app/app.routes.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
-import { LoginComponent } from './login.component';
-import { CreateAccountComponent } from './create-account.component';
-import { AccountComponent } from './account.component';
+import { LoginComponent } from './login/login.component';
+import { CreateAccountComponent } from './create-account/create-account.component';
+import { AccountComponent } from './account/account.component';
 
 export const routes: Routes = [
   { path: '', component: LoginComponent },
